Added docs. Added frontmatter support.

+++ commands/backup2github.py
@@ -1,5 +1,6 @@
 import click
 import csv
+import frontmatter
 import os
 import requests
 
@@ -17,8 +18,8 @@ def fetch_title(url):
     return title
 
 def sluggify(string):
-    string = string.replace(' ', '-')
     string = string.replace(':', '')
+    string = string.replace(' ', '-')
     string = string.lower()
 
     return string
@@ -30,19 +31,18 @@ def get_markdown_url(url):
 
     return url
 
+def makedirs(path):
+    if not os.path.exists(path):
+        os.makedirs(path)
 
 @click.command(context_settings=CONTEXT_SETTINGS)
 @click.option('--resource-list',
               required=True,
               help='CSV file listing the urls to backup.',
               )
-@click.option('--github-repo',
-              help='Repo in which to create backups. Ex: someorg/myrepo',
-              envvar='GITHUB_REPO',
-              )
 @click.option('--destination',
               required=True,
-              help='Local directory or github repo in which to write backups. Ex: path/to/dir, someorg/myrepo',
+              help='Local path or GitHub repo in which to write backups. Ex: path/to/dir, someorg/myrepo',
               envvar='BACKUP_DESTINATION',
               )
 @click.option('--github-token',
@@ -50,10 +50,35 @@ def get_markdown_url(url):
               envvar="GITHUB_API_TOKEN",
               )
 @common_params
-def backup2github(resource_list, github_token, github_repo, destination, yes, verbose, debug, noop):
-    """Backup a list of URLs to a GitHub repository.
+def backup2github(resource_list, github_token, destination, yes, verbose, debug, noop):
+    """Backup a list of URLs to a destination directory or GitHub repo.
 
-    Currently supports: HackMD.
+    The format of the --resource-list is expected to be a local CSV. Columns
+    are as follows:
+
+        - do_backup. If column is present, empty rows will not be backed up.
+
+        - resource_url. The URL to the file for backup. Supports raw files
+            directly.
+
+        - Any other columns will be ignored.
+
+    Currently supported destination is a local directory. GitHub repos coming
+    soon.
+
+    Currently supported backup URLs:
+
+        - any raw files
+
+        - HackMD web urls (fetched as markdown)
+
+    When processing markdown files, YAML frontmatter can be used for certain
+    keys:
+
+        - filename: Allows overriding of the filename, normally processed from
+            header. Ex: myfile.md
+
+        - path: Allows nesting the backup file in a subdirectory.
     """
     if debug: click.echo('>>> Debug mode: enabled')
     if noop: click.echo('>>> No-op mode: enabled (No operations affecting data will be run)')
@@ -94,10 +119,31 @@ def backup2github(resource_list, github_token, github_repo, destination, yes, ve
             url = get_markdown_url(url)
             r = requests.get(url)
             markdown = r.content.decode('utf-8')
-            click.echo('Writing file: ' + filepath)
-            with open(filepath, 'w') as f:
+            post = frontmatter.loads(markdown)
+
+            # Append source url to frontmatter.
+            post.metadata['source'] = row['resource_url']
+            markdown = frontmatter.dumps(post)
+
+            metadata = post.metadata
+            if metadata.get('filename'):
+                filename = metadata['filename']
+
+            if metadata.get('path'):
+                filepath = os.path.join(destination, metadata['path'])
                 if not noop:
+                    makedirs(filepath)
+            else:
+                filepath = destination
+
+            filepath = os.path.join(filepath, filename)
+
+            # TODO: Do a safety check to ensure we're not outside destination path.
+            click.echo('Writing file: ' + filepath)
+            if not noop:
+                with open(filepath, 'w') as f:
                     f.write(markdown)
+                    f.write("\n")
 
             # TODO: process hackmd-specific formatting.
